third_party/blink/web_tests/external/wpt/css/css-color/nested-color-mix-with-currentcolor-expected.txt

This is a testharness.js-based test.
[FAIL] Nested color-mix function with inner currentColor should inherit unresolved
  assert_equals: expected "lch(0 0 none)" but got "lch(0 0 0)"
Harness: the test ran to completion.
